RIA running but wrong
Disturbances are only a repeat of the 100-year fire composite raster for the RIA. Need yearly fires and harvesting rasters. Carbon is too high because growth curves are too hight. Limits will have to be set on growth curves and the clean-up and translation process will have to happen again. Saving this branch as is. Plotting was checked and fixed.